Now downloading checkpoint, also more markdown.
- Added some utilities inspired by the fastai notebooks.
- Ended up using wget instead of the HTTP library 'Just', as it wasn't
handling redirects nicely. It would write the contents of the redirect
landing page into the file to which it was redirecting. I thought I did
something wrong when I made the tar, but finally found that there was
html at the beginning of the file. wget works perfectly fine in this
scenario!
- Added a bit more prose to the last part of the notebook. Also improved
the subplot code to maximize screen real estate.